feat(): Add copy implementation (#1758)

+++ core/src/web/filesystem.ts
@@ -1,6 +1,8 @@
 import {WebPlugin} from './index';
 
 import {
+  CopyOptions,
+  CopyResult,
   FileAppendOptions,
   FileAppendResult,
   FileDeleteOptions,
@@ -15,10 +17,10 @@ import {
   GetUriResult,
   MkdirOptions,
   MkdirResult,
-  RenameOptions,
-  RenameResult,
   ReaddirOptions,
   ReaddirResult,
+  RenameOptions,
+  RenameResult,
   RmdirOptions,
   RmdirResult,
   StatOptions,
@@ -358,18 +360,45 @@ export class FilesystemPluginWeb extends WebPlugin implements FilesystemPlugin {
    * @return a promise that resolves with the rename result
    */
   async rename(options: RenameOptions): Promise<RenameResult> {
-    let {to, from, directory} = options;
+    return this._copy(options, true);
+  }
+
+  /**
+   * Copy a file or directory
+   * @param options the options for the copy operation
+   * @return a promise that resolves with the copy result
+   */
+  async copy(options: CopyOptions): Promise<CopyResult> {
+    return this._copy(options, false);
+  }
+
+  /**
+   * Function that can perform a copy or a rename
+   * @param options the options for the rename operation
+   * @param doRename whether to perform a rename or copy operation
+   * @return a promise that resolves with the result
+   */
+  private async _copy(options: CopyOptions, doRename: boolean = false): Promise<CopyResult> {
+    let {to, from, directory: fromDirectory, toDirectory} = options;
 
     if (!to || !from) {
       throw Error('Both to and from must be provided');
     }
 
+    // If no "to" directory is provided, use the "from" directory
+    if (!toDirectory) {
+      toDirectory = fromDirectory;
+    }
+
+    let fromPath = this.getPath(fromDirectory, from);
+    let toPath = this.getPath(toDirectory, to);
+
     // Test that the "to" and "from" locations are different
-    if (from === to) {
+    if (fromPath === toPath) {
       return {};
     }
 
-    if (to.startsWith(from)) {
+    if (toPath.startsWith(fromPath)) {
       throw Error('To path cannot contain the from path');
     }
 
@@ -378,7 +407,7 @@ export class FilesystemPluginWeb extends WebPlugin implements FilesystemPlugin {
     try {
       toObj = await this.stat({
         path: to,
-        directory
+        directory: toDirectory
       });
     } catch (e) {
       // To location does not exist, ensure the directory containing "to" location exists and is a directory
@@ -390,7 +419,7 @@ export class FilesystemPluginWeb extends WebPlugin implements FilesystemPlugin {
       if (toPathComponents.length > 0) {
         let toParentDirectory = await this.stat({
           path: toPath,
-          directory,
+          directory: toDirectory,
         });
 
         if (toParentDirectory.type !== 'directory') {
@@ -407,12 +436,12 @@ export class FilesystemPluginWeb extends WebPlugin implements FilesystemPlugin {
     // Ensure the "from" object exists
     let fromObj = await this.stat({
       path: from,
-      directory: directory
+      directory: fromDirectory,
     });
 
     // Set the mtime/ctime of the supplied path
     let updateTime = async (path: string, ctime: number, mtime: number) => {
-      let fullPath: string = this.getPath(directory, path);
+      let fullPath: string = this.getPath(toDirectory, path);
       let entry = await this.dbRequest('get', [fullPath]) as EntryObj;
       entry.ctime = ctime;
       entry.mtime = mtime;
@@ -425,24 +454,28 @@ export class FilesystemPluginWeb extends WebPlugin implements FilesystemPlugin {
         // Read the file
         let file = await this.readFile({
           path: from,
-          directory
+          directory: fromDirectory
         });
 
-        // Remove the file
-        await this.deleteFile({
-          path: from,
-          directory
-        });
+        // Optionally remove the file
+        if (doRename) {
+          await this.deleteFile({
+            path: from,
+            directory: fromDirectory
+          });
+        }
 
         // Write the file to the new location
         await this.writeFile({
           path: to,
-          directory,
+          directory: toDirectory,
           data: file.data
         });
 
-        // Copy the mtime/ctime of the original file
-        await updateTime(to, fromObj.ctime, fromObj.mtime);
+        // Copy the mtime/ctime of a renamed file
+        if (doRename) {
+          await updateTime(to, fromObj.ctime, fromObj.mtime);
+        }
 
         // Resolve promise
         return {};
@@ -456,35 +489,40 @@ export class FilesystemPluginWeb extends WebPlugin implements FilesystemPlugin {
           // Create the to directory
           await this.mkdir({
             path: to,
-            directory,
+            directory: toDirectory,
             createIntermediateDirectories: false,
           });
 
-          // Copy the mtime/ctime of the original directory
-          await updateTime(to, fromObj.ctime, fromObj.mtime);
+          // Copy the mtime/ctime of a renamed directory
+          if (doRename) {
+            await updateTime(to, fromObj.ctime, fromObj.mtime);
+          }
         } catch (e) {
         }
 
         // Iterate over the contents of the from location
         let contents = (await this.readdir({
           path: from,
-          directory
+          directory: fromDirectory,
         })).files;
 
         for (let filename of contents) {
           // Move item from the from directory to the to directory
-          await this.rename({
+          await this._copy({
             from: `${from}/${filename}`,
             to: `${to}/${filename}`,
-            directory,
-          });
+            directory: fromDirectory,
+            toDirectory,
+          }, doRename);
         }
 
-        // Remove the original from directory
-        await this.rmdir({
-          path: from,
-          directory
-        });
+        // Optionally remove the original from directory
+        if (doRename) {
+          await this.rmdir({
+            path: from,
+            directory: fromDirectory
+          });
+        }
     }
 
     return {};
